Remove include hdf.f90 statements
Recent versions of the HDF 4 library removed the hdf.f90 'header' file
that we used to include parameters. The Fortran 77 version of the file,
hdf.inc, remains but cannot be directly included within Fortran 90 code.
I got around this by generating the desired header from the remaining
one using sed but that is a non-obvious step that prevents straight-
forward use of an external library.

This commit creates a common module from hdf.inc and replaces the
include commands with uses of that module. This should be robust against
future revision of the HDF library.